backend/requirements.txt
django>=6.0,<7.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0
Pillow>=10.0
python-decouple>=3.8
psycopg2-binary>=2.9
requests>=2.31
celery>=5.3
redis>=5.0
gunicorn>=21.2
